requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2024.6.1
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==2.1.5
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.0.0
propcache==0.3.1
psutil==7.0.0
py-cpuinfo==9.0.0
PyJWT==2.10.1
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.15.3
six==1.17.0
sympy==1.13.3
torch==2.7.0+cu128
torchaudio==2.7.0+cu128
torchvision==0.22.0+cu128
tqdm==4.67.1
twilio==9.6.1
typing_extensions==4.12.2
tzdata==2025.2
ultralytics==8.3.142
ultralytics-thop==2.0.14
urllib3==2.4.0
yarl==1.20.0
